+++ Actions/UnetGetPeerType.cs
@@ -100,4 +100,3 @@ namespace HutongGames.PlayMaker.Ecosystem.Networking.Actions
 		}
 	}
 }
-